feat(budget-management): 新增预算费用模块
- 添加预算费用相关实体、Mapper、Service及Controller- 实现预算费用的分页查询、单条查询、新增、修改和删除功能
- 新增批量修改状态功能
- 优化查询结果，增加预算科目编码和名称等关联信息

+++ src/main/java/org/psd/budget_management/controller/BudgetItemLogController.java
@@ -90,8 +90,7 @@ public class BudgetItemLogController {
      * @return 删除结果
      */
     @DeleteMapping
-    public Result deleteByIds(@RequestParam("idList") List<Long> idList) {
+    public Result deleteByIds(@RequestParam("idList") List<Integer> idList) {
         return new Result(200, "执行成功", this.budgetItemLogService.removeByIds(idList));
     }
 }
-
